requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.6.1
httpx==0.27.2
orjson==3.10.15
python-multipart==0.0.9
python-dotenv>=1.0.0
defusedxml>=0.7.1

psycopg2-binary>=2.9.9
